+++ s15-10/demo/src/main/java/s1510/demo/repository/MatchRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import s1510.demo.model.Match;
 
 @Repository
-public interface MatchRepository extends GenericRepo<Match, Integer> {
+public interface MatchRepository extends GenericRepo<Match, Long> {
 }
